fix some shady filesystem code

+++ src/game/g_scr_main_mp.cpp
@@ -3250,7 +3250,7 @@ void GScr_OpenFile()
 
 			if ( len >= 0 )
 			{
-				level.openScriptIOFileBuffers[filenum] = (char *)Z_MallocInternal(len + sizeof(int));
+				level.openScriptIOFileBuffers[filenum] = (char *)Z_MallocInternal(len + sizeof(intptr_t));
 				FS_Read(level.openScriptIOFileBuffers[filenum], len, tempFile);
 				FS_FCloseFile(tempFile);
 				level.openScriptIOFileBuffers[filenum][len] = 0;
